Safari-611.1.10/LayoutTests/inspector/audit/version-expected.txt

Tests Audit.VERSION value.


== Running test suite: Audit.VERSION
-- Running test case: Audit.VERSION.MatchesFrontend
PASS: The audit system version should match the frontend version.
